function delete products and verifications

+++ src/product/product.controller.js
@@ -5,6 +5,8 @@ import ProductMongo from "./product.DAO.js"
 import CustomErrors from '../tools/customError.js'
 import { createProductErrorInfo } from '../tools/info.js'
 import EErrors from '../tools/EErrors.js'
+import { userService } from '../user/user.controller.js'
+import { ShowProductDto } from './product.DTO.js'
 
 export const productService = new ProductRepository(new ProductMongo())
 
@@ -33,9 +35,9 @@ export const getProductById = async(req, res) => {
 
 export const addProduct = async(req, res) => {
     try {
-        const {title, category, status, description, price, code, stock} = req.body;
+        const {title, category, description, price, stock} = req.body;
     
-        if (title === undefined || category === undefined || status === undefined || description === undefined || price === undefined || code === undefined || stock === undefined){
+        if (title === undefined || category === undefined || description === undefined || price === undefined || stock === undefined){
             CustomErrors.createError({
                 name: 'Error al crear el producto',
                 message: 'Debes enviar todos los campos requeridos',
@@ -43,14 +45,17 @@ export const addProduct = async(req, res) => {
                 code: EErrors.INVALID_TYPE
             })
         }
-        await productService.addProduct(req.body)
+        const p = new ShowProductDto(req.body).getProduct()
+        //SOLO RECIBE USUARIOS PREMIUM
+        p.owner = req.session.user.role === 'premium'? req.session.user.email : 'admin'
+
+        await productService.addProduct(p)
         const products = await productService.getProducts()
         io.emit('products', products)//<--envia al socket
-        res.status(201).send(req.body)
-
+        return res.redirect('/')
     } catch (err) {
-        req.logger.error(`${err.name}: ${err.message}${err.cause}`)
-        res.status(400).send(err)
+        req.logger.warning(`${err.name}: ${err.message}${err.cause}`)
+        res.render('createProduct', {messageError: 'Debes enviar todos los campos requeridos'})
     }
 }
 
@@ -70,15 +75,54 @@ export const updateProduct = async(req, res) => {
     }
 }
 
-export const deleteProduct = async(req, res) => {
+export const deleteProduct2 = async(req, res) => {
     try {
-        await productService.deleteProduct(req.params.pid)
-        const products = await productService.getProducts()
-        io.emit('products', products)//<--envia al socket
-        req.logger.info(`se eliminó correctamente`)
-        res.status(204).send({status: `se elimino correctamente`});
+        const user = await userService.getByEmail(req.session.user.email)
+        const product = await productService.getProductById(req.params.pid)
+        req.logger.debug('1')
+        if (user.role === 'premium' && user.email == product.owner) {
+            // await productService.deleteProduct(req.params.pid)
+            req.logger.debug('2')
+            req.logger.info(`se eliminó correctamente`)
+            return res.send('ok')
+            // return res.status(302).redirect('/')
+            // const products = await productService.getProducts()
+            // io.emit('products', products)//<--envia al socket
+        }
+        
+        if (user.role === 'admin') {
+            // await productService.deleteProduct(req.params.pid)
+            req.logger.debug('2')
+            req.logger.info(`se eliminó correctamente`)
+            return res.send('ok')
+            // return res.status(302).redirect('/')
+            // const products = await productService.getProducts()
+            // io.emit('products', products)//<--envia al socket
+        }
+        
+        req.logger.debug('3')
+        req.logger.info('No estas autorizado a eliminar este producto')
+        return res.status(400).send(`Esta operacion no se puede hacer`);
     }
     catch (err) {
         res.status(500).send({error: err})
+    }
+}
+
+export const deleteProduct = async (req, res) => {
+    try {
+        const product = await productService.getProductById(req.params.pid)
+        const user = await userService.getByEmail(req.user.user.email)
+
+        if (user.role === 'premium' && user.email !== product.owner) {
+            return req.logger.info(`no puedes borrar este producto`)
+        }
+        
+        await productService.deleteProduct(req.params.pid)
+        req.logger.info(`se eliminó correctamente`)
+        return res.status(200).send('Se eliminó el producto!')
+    }
+    catch (err) {
+        console.log(err);
     }
 }
